+++ server/social.js
@@ -1,5 +1,11 @@
 const fetch = require( 'node-fetch' );
 
+
+
+// fetch( 'https://twitter.com/elonmusk' ).then( res => res.text() ).then( r => console.log( r ) );
+
+
+
 module.exports = function ( app ) {
     app.get( '/social/google/trends', ( req, res ) => {
         fetch( "https://trends.google.com/trends/trendingsearches/daily/rss?geo=US" )